backend/requirements.txt
Django
psycopg2-binary
requests
djangorestframework
djangorestframework-simplejwt